Read overview history separately and group preview rows (DRU-516)
Two statements replace the cross join of the history aggregate into the previews, so the empty-row guard goes. Each section is built once from its grouped rows. The overview tests join test_dashboard.py. An empty app filter reads all apps.
