+++ backend/druks/api/dashboard.py
@@ -1,5 +1,8 @@
+from itertools import groupby
+from operator import attrgetter
+
 from fastapi import APIRouter, HTTPException, Response
-from sqlalchemy import case, func, select, true
+from sqlalchemy import case, func, select
 
 from druks.api.schemas import (
     DashboardOverview,
@@ -17,6 +20,7 @@ from druks.events.models import Event
 from druks.settings import load_settings
 
 PAGE_SIZE = 200
+PREVIEW_SIZE = 4
 router = APIRouter(prefix="/api/dashboard", tags=["dashboard"])
 
 
@@ -33,7 +37,7 @@ async def get_overview(response: Response, app: str | None = None) -> DashboardO
     owners = {
         workflow.kind: owner.name for owner in apps.values() for workflow in owner.workflows()
     }
-    current = (
+    current_runs = (
         Run.get_open_subjects(
             kinds=list(owners),
             states=(RunState.PARKED, RunState.RUNNING, RunState.FAILED, RunState.ORPHANED),
@@ -43,15 +47,19 @@ async def get_overview(response: Response, app: str | None = None) -> DashboardO
         .subquery()
     )
     section = case(
-        (current.c.state == RunState.PARKED, "needs_you"),
-        (current.c.state == RunState.RUNNING, "running"),
+        (current_runs.c.state == RunState.PARKED, "needs_you"),
+        (current_runs.c.state == RunState.RUNNING, "running"),
         else_="failed",
     )
-    request_time = case((current.c.state == RunState.PARKED, current.c.input_requested_at))
-    request_id = case((current.c.state == RunState.PARKED, current.c.run_id))
-    ranked = (
+    # Both request columns are NULL outside needs_you, so running and failed
+    # order by updated_at alone.
+    request_time = case(
+        (current_runs.c.state == RunState.PARKED, current_runs.c.input_requested_at)
+    )
+    request_id = case((current_runs.c.state == RunState.PARKED, current_runs.c.run_id))
+    ranked_runs = (
         select(
-            current,
+            current_runs,
             section.label("section"),
             func.count().over(partition_by=section).label("total"),
             func.row_number()
@@ -60,70 +68,62 @@ async def get_overview(response: Response, app: str | None = None) -> DashboardO
                 order_by=(
                     request_time.asc(),
                     request_id.asc(),
-                    current.c.updated_at.desc(),
-                    current.c.run_id.desc(),
+                    current_runs.c.updated_at.desc(),
+                    current_runs.c.run_id.desc(),
                 ),
             )
             .label("position"),
         )
         .where(
-            (current.c.state != RunState.PARKED)
-            | (current.c.input_requested_at.is_not(None) & (current.c.presentation != ""))
+            (current_runs.c.state != RunState.PARKED)
+            | (current_runs.c.input_requested_at.is_not(None) & (current_runs.c.presentation != ""))
         )
         .subquery()
     )
-    previews = select(ranked).where(ranked.c.position <= 4).subquery()
-    history = (
+    previews = (
         select(
-            func.max(Event.created_at)
-            .filter(Event.type == "workflow.finished")
-            .label("last_finished_at"),
-            func.max(Event.created_at)
-            .filter(Event.type == "workflow.failed")
-            .label("last_failed_at"),
-        )
-        .where(Event.app.in_(apps), Event.type.in_(("workflow.finished", "workflow.failed")))
-        .subquery()
-    )
-    statement = (
-        select(
-            history,
-            previews.c.section,
-            previews.c.total,
-            previews.c.run_id.label("run"),
-            previews.c.kind,
-            previews.c.state,
-            previews.c.subject_type,
-            previews.c.subject_id,
-            func.left(previews.c.subject_label, 240).label("subject_label"),
-            previews.c.updated_at,
-            previews.c.input_requested_at.label("parked_at"),
-            previews.c.request_label,
+            ranked_runs.c.section,
+            ranked_runs.c.total,
+            ranked_runs.c.run_id.label("run"),
+            ranked_runs.c.kind,
+            ranked_runs.c.state,
+            ranked_runs.c.subject_type,
+            ranked_runs.c.subject_id,
+            func.left(ranked_runs.c.subject_label, 240).label("subject_label"),
+            ranked_runs.c.updated_at,
+            ranked_runs.c.input_requested_at.label("parked_at"),
+            ranked_runs.c.request_label,
             func.left(Artifact.title, 240).label("artifact_title"),
-            previews.c.presentation,
-            previews.c.request_url,
-            func.left(previews.c.failure, 2048).label("failure"),
+            ranked_runs.c.presentation,
+            ranked_runs.c.request_url,
+            func.left(ranked_runs.c.failure, 2048).label("failure"),
         )
-        .select_from(history)
-        .outerjoin(previews, true())
-        .outerjoin(Artifact, Artifact.agent_call_id == previews.c.latest_call_id)
-        .order_by(previews.c.section, previews.c.position)
+        .outerjoin(Artifact, Artifact.agent_call_id == ranked_runs.c.latest_call_id)
+        .where(ranked_runs.c.position <= PREVIEW_SIZE)
+        .order_by(ranked_runs.c.section, ranked_runs.c.position)
     )
-    rows = (await db_session().execute(statement)).all()
+    history = select(
+        func.max(Event.created_at)
+        .filter(Event.type == "workflow.finished")
+        .label("last_finished_at"),
+        func.max(Event.created_at).filter(Event.type == "workflow.failed").label("last_failed_at"),
+    ).where(Event.app.in_(list(apps)), Event.type.in_(("workflow.finished", "workflow.failed")))
+    session = db_session()
     sections = {
         name: DashboardSection(total=0, rows=[]) for name in ("needs_you", "running", "failed")
     }
-
-    for row in rows:
-        if row.run:
-            sections[row.section].total = row.total
-            sections[row.section].rows.append(
+    for name, group in groupby((await session.execute(previews)).all(), key=attrgetter("section")):
+        preview = list(group)
+        sections[name] = DashboardSection(
+            total=preview[0].total,
+            rows=[
                 DashboardRun.model_validate({**row._mapping, "app": owners[row.kind]})
-            )
+                for row in preview
+            ],
+        )
+    times = (await session.execute(history)).one()
     return DashboardOverview(
-        **sections,
-        last_finished_at=rows[0].last_finished_at,
-        last_failed_at=rows[0].last_failed_at,
+        **sections, last_finished_at=times.last_finished_at, last_failed_at=times.last_failed_at
     )
